requirements.txt
PyMuPDF
python-dotenv
phonenumbers
python-magic
getch
